refactor: move SearchBar component to its own search directory
