+++ frontend/src/components/layout/Header.tsx
@@ -6,7 +6,7 @@ import Link from 'next/link'
 import { useRouter } from 'next/navigation'
 import { User as UserIcon, Star as StarIcon } from 'lucide-react'
 import { useAuth } from '@/context/AuthContext'
-import SearchBar from '@/components/common/SearchBar'
+import SearchBar from '@/components/search/SearchBar'
 import HighlightedText from '@/components/common/HighlightedText'
 import Dropdown from '@/components/common/Dropdown'
