+++ frontend/src/components/layout/Hero.tsx
@@ -1,7 +1,7 @@
 // src/components/Hero.tsx
 
 import HighlightedText from '@/components/common/HighlightedText'
-import SearchBar from '@/components/common/SearchBar'
+import SearchBar from '@/components/search/SearchBar'
 
 const Hero = () => {
   return (
